+++ ogcapi-stable/ogcapi-html/src/main/javascript/src/components/MapLibre/stories.jsx
@@ -92,10 +92,6 @@ Plain.args = {
               id: "transportationground",
               layers: ["transportationground"],
             },
-            {
-              id: "transportationground",
-              layers: ["transportationground"],
-            },
           ],
         },
       ],
